+++ dev/include/modulator.h
@@ -2,11 +2,11 @@
 #define MODULATION_H
 #include "common.h"
 
-vector<complex<double>> modulator(vector<int16_t> bits, int len_bits, string type);
-vector<int16_t> demodulator(const vector<complex<double>>& symbols, string type);
+vector<complex<float>> modulator(vector<int16_t> bits, int len_bits, string type);
+vector<int16_t> demodulator(const vector<complex<float>>& symbols, string type);
 
-vector<complex<double>> generate_header(size_t size, SharedData &sd);
-uint16_t decode_header(const vector<complex<double>> signal, SharedData &sd);
+vector<complex<float>> generate_header(size_t size, SharedData &sd);
+uint16_t decode_header(const vector<complex<float>> signal, SharedData &sd);
 
 int bits_per_symbol(string type);
